api/requirements.txt
fastapi==0.97.0
httpx==0.24.1
pydantic==1.10.8
python-dotenv==1.0.0
uvicorn==0.22.0
asyncpg==0.27.0
